doubles

+++ src/main/kotlin/com/example/cache/InMemoryModel.kt
@@ -3,9 +3,9 @@ package com.example.cache
 import kotlinx.serialization.Serializable
 
 object InMemoryModel {
-    val X = arrayListOf<Int>(0) // arrayListOf(1, 2, 3, 4, 5, 25)
-    val Y = arrayListOf<Int>(0) // arrayListOf(1, 2, 4, 8, 16, 25)
-    val Z = arrayListOf<Int>(0) // arrayListOf(1, 2, 4, 8, 16, 25)
+    val X = arrayListOf<Double>(0.0) // arrayListOf(1, 2, 3, 4, 5, 25)
+    val Y = arrayListOf<Double>(0.0) // arrayListOf(1, 2, 4, 8, 16, 25)
+    val Z = arrayListOf<Double>(0.0) // arrayListOf(1, 2, 4, 8, 16, 25)
 }
 
 fun InMemoryModel.toRemote(): ModelRemote {
@@ -14,9 +14,9 @@ fun InMemoryModel.toRemote(): ModelRemote {
 
 @Serializable
 data class ModelRemote(
-    val x: List<Int> = emptyList(),
-    val y: List<Int> = emptyList(),
-    val z: List<Int> = emptyList()
+    val x: List<Double> = emptyList(),
+    val y: List<Double> = emptyList(),
+    val z: List<Double> = emptyList()
 )
 
 @Serializable
